* ens-sync-script-healthcheck: (39 commits)
  codereview changerequest changes
  release(network-contracts): v6.1.0
  release(config): @streamr/config 4.0.5
  ETH-585 docker images build and test in CI (#616)
  Eth 596 subgraph fixeschanges (#631)
  fix: merge hazard from flagWithMetadata PR
  cleanup: reorder setParam function, uint256 -> uint in Policies
  feat: make slashing fraction configurable (#624)
  rename flagWithMetadata -> flag, and remove the previous implementation (#630)
  release(config): @streamr/config 4.0.4
  streamr 1.0 mumbai deployment addresses update
  Modularize operator (#629)
  doc: add clean step to push-to-docker-hub instructions
  release(network-contracts): 6.0.1
  fix: remove accidentally duplicated function
  Eth 584 erc677 in operator factory (#627)
  added hardhat-contract-sizer to contract build process
  ETH-579: Check Operator contract function guards (#628)
  deoc: added addresses preloaded with DATA to fastchain readme
  release(config): @streamr/config 4.0.3
  ...

# Conflicts:
#	.github/workflows/ci.yaml
#	packages/network-contracts/package.json

+++ packages/network-contracts/contracts/OperatorTokenomics/OperatorPolicies/IUndelegationPolicy.sol
@@ -3,6 +3,8 @@
 pragma solidity ^0.8.13;
 
 interface IUndelegationPolicy {
-    function setParam(uint256 param) external;
-    function onUndelegate(address delegator, uint256 amount) external;
+    function setParam(uint param) external;
+
+    /** can throw to prevent undelegation. Gets called BEFORE entering the undelegation queue */
+    function onUndelegate(address delegator, uint amount) external;
 }
