+++ packages/network-contracts/contracts/OperatorTokenomics/SponsorshipPolicies/ILeavePolicy.sol
@@ -3,6 +3,6 @@
 pragma solidity ^0.8.13;
 
 interface ILeavePolicy {
-    function setParam(uint256 param) external;
+    function setParam(uint param) external;
     function getLeavePenaltyWei(address operator) external view returns (uint leavePenaltyWei);
 }
